+++ src/Console/Commands/CreateUsersCommand.php
@@ -6,12 +6,13 @@ namespace Patrikjak\Auth\Console\Commands;
 
 use Illuminate\Console\Command;
 use Illuminate\Contracts\Config\Repository as Config;
+use Illuminate\Hashing\HashManager;
 use Patrikjak\Auth\Exceptions\RoleNotFoundException;
 use Patrikjak\Auth\Factories\UserFactory;
 use Patrikjak\Auth\Models\Role;
 use Patrikjak\Auth\Models\User;
-use Patrikjak\Auth\Repositories\Interfaces\RoleRepository;
-use Patrikjak\Auth\Repositories\Interfaces\UserRepository;
+use Patrikjak\Auth\Repositories\Contracts\RoleRepository;
+use Patrikjak\Auth\Repositories\Contracts\UserRepository;
 
 final class CreateUsersCommand extends Command
 {
@@ -27,18 +28,22 @@ final class CreateUsersCommand extends Command
      */
     protected $description = 'Create users';
 
-    public function handle(Config $config, RoleRepository $roleRepository, UserRepository $userRepository): void
-    {
-        $user = $this->getUserData($config, $roleRepository);
+    public function handle(
+        Config $config,
+        HashManager $hashManager,
+        RoleRepository $roleRepository,
+        UserRepository $userRepository,
+    ): void {
+        $user = $this->getUserData($config, $hashManager, $roleRepository);
         $userRepository->createAndReturnUser($user);
 
         while ($this->confirm('Do you want to create another user?')) {
-            $user = $this->getUserData($config, $roleRepository);
+            $user = $this->getUserData($config, $hashManager, $roleRepository);
             $userRepository->createAndReturnUser($user);
         }
     }
 
-    private function getUserData(Config $config, RoleRepository $roleRepository): User
+    private function getUserData(Config $config, HashManager $hashManager, RoleRepository $roleRepository): User
     {
         $defaultPassword = $config->get('pjauth.user_default_password');
         $defaultRoleSlug = $config->get('pjauth.default_role_slug', 'superadmin');
@@ -63,9 +68,10 @@ final class CreateUsersCommand extends Command
         }
 
         $userModel = UserFactory::getUserModel();
+
         $userModel->name = $name;
         $userModel->email = $email;
-        $userModel->password = $password;
+        $userModel->password = $hashManager->make($password);
         $userModel->role_id = $role->id;
 
         return $userModel;