+++ src/Listeners/DeleteRegisterInviteListener.php
@@ -5,16 +5,16 @@ declare(strict_types=1);
 namespace Patrikjak\Auth\Listeners;
 
 use Patrikjak\Auth\Events\RegisteredViaInviteEvent;
-use Patrikjak\Auth\Repositories\Interfaces\UserRepository;
+use Patrikjak\Auth\Repositories\Contracts\RegisterInviteRepository;
 
 readonly class DeleteRegisterInviteListener
 {
-    public function __construct(private UserRepository $userRepository)
+    public function __construct(private RegisterInviteRepository $registerInviteRepository)
     {
     }
 
     public function handle(RegisteredViaInviteEvent $event): void
     {
-        $this->userRepository->deleteRegisterInvite($event->user->email);
+        $this->registerInviteRepository->delete($event->user->email);
     }
 }