+++ src/Repositories/Contracts/RoleRepository.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Patrikjak\Auth\Repositories\Interfaces;
+namespace Patrikjak\Auth\Repositories\Contracts;
 
 use Illuminate\Support\Collection;
 use Patrikjak\Auth\Models\Role;